python/uploads/requirements.txt
alembic==1.13.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
gevent==23.9.1
greenlet==3.0.3
PyJWT==2.8.0
redis==5.0.1
PyMySQL==1.1.1
PyMySQL[rsa]
SQLAlchemy==2.0.31